# Conflicts:
#	src/business/DTO/ProjectDTO.java
#	src/business/UCC/ViewOptionUCC.java
#	src/business/UCC/ViewOptionUCCImpl.java
#	src/persistence/ProjectDAO.java
#	src/persistence/ProjectDAOImpl.java
#	src/utilities/Utility.java
#	src/view/dashboard/DashboardController.java
#	src/view/dashboard/ViewOptionController.java
#	src/view/dashboard/dashboard.fxml

+++ src/business/shape/Coordinates.java
@@ -6,34 +6,33 @@ import java.io.Serializable;
  * Pair of coordinates, with their respective units and options.
  * only work with default units for now (cm).
  * https://en.wikibooks.org/wiki/LaTeX/PGF/TikZ#Specifying_Coordinates
- *
  */
 public class Coordinates implements Serializable {
-    private float x = 0;
-    private float y = 0;
+    private double x = 0;
+    private double y = 0;
 
-    public Coordinates(float x, float y) {
+    public Coordinates(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public String print(){
-        return "("+this.x+","+this.y+") ";
+    public String print() {
+        return "(" + this.x + "," + this.y + ") ";
     }
 }
